backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.7.0
pydantic-settings>=2.0.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles>=23.0.0
chromadb-client>=1.1.1
onnxruntime>=1.23.0
tokenizers>=0.20.0
groq>=0.9.0
b2sdk==1.26.0
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
langchain-text-splitters>=0.2.0
